Migrates CurrencyUtils formatSingle to Settings
